A fix for this issue : http://code.google.com/p/fsnet/issues/detail?id=257 .
However i found an other issue when i tested subscribe/unsubscribe event. Please read the new issue comment. 

git-svn-id: http://fsnet.googlecode.com/svn@1791 07c314b0-eee8-11dd-b478-27b5dcd53177

+++ trunk/facade/src/main/java/fr/univartois/ili/fsnet/facade/InteractionRoleFacade.java
@@ -53,9 +53,11 @@ public class InteractionRoleFacade {
         interactionRolePK.setInteractionId(i.getId());
         interactionRolePK.setSocialEntityId(se.getId());
         InteractionRole interactionRole = em.find(InteractionRole.class, interactionRolePK);
-        se.getRolesInInteractions().remove(interactionRole);
-        i.getRoles().remove(interactionRole);
-        em.remove(interactionRole);
+        if(interactionRole != null){
+	        se.getRolesInInteractions().remove(interactionRole);
+	        i.getRoles().remove(interactionRole);
+	        em.remove(interactionRole);
+        }
     }
 
     /**
